cau 1/requirements.txt
mysql-connector-python==8.0.33
flask==2.0.1
flask-cors==3.0.10 